requirements.txt
# Core Math & Data
numpy==1.21.5
matplotlib==3.6.1
tqdm
scipy

# Configuration
hydra-core>=1.3.0
omegaconf

# Deep Learning (PyTorch - CPU is fine for parsing, but install CUDA if you have it)
torch>=2.0.0
torchvision

# Waymo Open Dataset & TensorFlow (Required for parsing .tfrecords)
# We use the CPU version of TF to keep the environment lighter for this phase
tensorflow-cpu==2.11.0
waymo-open-dataset-tf-2-11-0

# Protocol Buffers - STRICT PINNING REQUIRED
# Waymo's pre-compiled protos often break with Protobuf 4.x/5.x
protobuf==3.19.6